+++ CHANGELOG.md
@@ -4,8 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
-## [Unreleased]
+## [0.4.0] - 2023-10-13
+- Added strum EnumString to PacketKind to allow converting to/from a string
 - Added method for writing a string prefixed with a u8 length byte
+- Updated GAME_IDENTIFIER to include a name
+- Changed lookup functions to return an owned value
 
 ## [0.3.0] - 2023-09-28
 - Added unit and integration tests (in-progress)